requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
python-multipart==0.0.9
langchain==0.3.6
langchain-community==0.3.5
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
pypdf==5.0.1
python-pptx==0.6.23
beautifulsoup4==4.12.3
requests==2.32.3
tavily-python==0.3.6
tiktoken==0.7.0
pydantic==2.9.2
groq==0.32.0
